Split custom generator into 2 classes.
This is to extract the method of generating a class via an array of
strings from the custom generator. The new baseclass CGeneratorStringReplace
now produces strct* files bases on an abstract array of const char *
returned by the abstract getter members. The former custom file takes
care of loading the sources and overwrites the getters to return the
proper source string.

This way we can base the inbuild geration using precompiled string on the
very same baseclass CGeneratorStringReplace.