add generated lexical.cxx to git in order to compile on windows, without the lex and yacc tools
On windows lex and yacc (flex/bison) are not easily available. As they are
not changed in a normal build process, they can be added to the git repo.
As long as the parser generater input files are not changed this will at
least compile on windows, making it possible to use the tool. For further
development of the tool flex/bison might be necessary at some point.